requirements.txt
mcp[cli]
langchain
langchain-mcp-adapters
langchain-google-genai
langgraph
pypdf
python-dotenv
